pyproject.toml
[tool.poetry]
name = "rag-chat"
version = "0.1.0"
description = "AI Conversation system with RAG to interact with structured and unstructured data."
authors = ["BennyDanielT <benny28dany@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
streamlit = "^1.33.0"
python-dotenv = "^1.0.1"
langchain = "^0.1.15"
pypdf2 = "^3.0.1"
pyprojroot = "^0.3.0"
mypy = "^1.9.0"
faiss-cpu = "^1.8.0"
langchain-google-genai = "^1.0.2"
google-generativeai = "^0.5.0"
langchain-chroma = { version = "^0.1.0", python = ">=3.12,<3.13" }
langchain-openai = "^0.1.3"
langchain-mistralai = "^0.1.2"
pypdf = "^4.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
